First github hookup.
Should really git from the start in future.

The current implementation provides the functionality to create
a timeline container, within which it populates a spine and interval
markers. It animates the extension of the spine and the introduction
of the markers.

Functionality is also there to map a date to a location, supposedly
making the creation of moments much easier than they would be
otherwise. Everything's done in coffeescript which is a first
attempt but so far it's quite nice. Old js implementation in the
timeline.js.old probably not even worth looking at.

TODO - Have to get the moments working
        - First step, generate test data
        - Second, dot the event horizons on the spine
        - Third, work on the interface for extending the
          info outward. General idea shown below
     - Partner this with a CATe event scraper, probably
       best to do in the classy-cate fork rather than this
       specific repo
     - Work on adding in customisations to the animation,
       interval distance etc. Make everything a bit more
       robust and ask someone how they'd tidy up the whole
       hot potato thing with the container variable.
       Ideally need to store the container jQuery object
       in a larger scope, just not quite globally...