Made fix to misalignment
Need to get some sort of global variable system working.

Caused a problem because of the way that I was calculating the
pctPerInterval value. Made an alteration to push the last interval to
the very end of the line and that got really nasty due to no change in
the container.data object.

Need a convinient way of accessing the data jquery object, otherwise
stuff like this is going to happen all over again. Wasted 20 minutes.