Implemented working layering.
Layering seems to work relatively well now, so there's a relief.
Update CSS now deals with the due date wires as well as the movement
of the actual labels. Quite happy with the result actually, better
than I expected it to be. Now just need to sort out implementing
links in the data structure and an expand all functionality on the
origin circle.