Implemented expand/collapse all.
The function works! Had a bit of an issue where I didn't realise
that calling the click trigger on all the moments was causing each
height recalculation to occur on completion of the previous, ie
the animations kept stacking up. Working quite nicely now though,
and is actually starting to look like this might not be a pile of shit!

Must be the music- From Me to You, John Lewis & Friends.
  <3 my iTunes.