Added in the real Reasoning data.
So now that's in you can see, in collapsed the layering is sort of
acceptable whereas when expanded things should tesselate more
effectively. Think this is more about the clashing functions than the
algorithm, cause if there's space it should sort of fit. Either that or
I should run it twice just to see if anything can fill the space, seems
like for effectiveness that should be a recursion though.

NB - recusion cause it should have a termination for when it gets to
the most optimal filling. Ummm, give it some consideration.

TODO - animate the extension of the container height.