Meet version 1.0
Sorted everything out so it works relatively ok. Even the colour scheme
doesn't seem to be too bad, but that's my opinion.

Works with classy-cate pretty nicely, though clearly the div animation
is rather expensive for so many timelines. May have to think about
binding their creation to the title of the module, or having a selective
view of the different timelines. That would avoid everything being
processed at once though obviously it's rather a stopgap.

TODO - Fix the clashing bug. Feeling is that rightmost is the culprit,
       I'll have to have a look later though.
     - Refactor, well, everything. It's messy and expensive where it
       doesn't need to be- sort it out.
     - Work on a better algorithm for the layering process. Idealy each
       element will position itself in the best gap. Probably something
       like layering the labels with the greatest widths first, rather
       than just a left to right motion.
     - Update the README so people might actually have a chance of using
       the damn thing.

BUGS - As mentioned in todo's, there's one with clashing. Double check
       the rightmost function and go from there.